# crm-dashboard/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Church Management System API
  description: |
    API for managing church members, cell groups (CGF), attendance, ministries, and events.
    This API provides CRUD operations for all entities and dashboard analytics endpoints.
  version: 1.0.0
  contact:
    name: Church Admin
    email: admin@church.org
  license:
    name: MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
  - url: https://api.church.org/v1
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /attendance:
    get:
      tags: [Attendance]
      summary: Get all attendance records
      description: Retrieve attendance records with filtering options
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: no_jemaat
          in: query
          description: Filter by member ID
          schema:
            type: integer
        - name: cg_id
          in: query
          description: Filter by CGF ID
          schema:
            type: string
        - name: tanggal
          in: query
          description: Filter by date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: keterangan
          in: query
          description: Filter by attendance status
          schema:
            type: string
            enum: [hadir, izin, tidak_hadir, tamu]
        - name: start_date
          in: query
          description: Start date for date range filter
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: End date for date range filter
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceListResponse'
              example:
                success: true
                data:
                  - id: 1
                    no_jemaat: 1
                    cg_id: "CGF01"
                    tanggal: "2024-01-15"
                    keterangan: "hadir"
                    member:
                      nama_jemaat: "John Doe"
                pagination:
                  page: 1
                  limit: 10
                  total: 500
                  totalPages: 50

    post:
      tags: [Attendance]
      summary: Record attendance
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceCreate'
      responses:
        '201':
          description: Attendance recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /attendance/bulk:
    post:
      tags: [Attendance]
      summary: Bulk record attendance
      description: Record attendance for multiple members at once
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cg_id, tanggal, records]
              properties:
                cg_id:
                  type: string
                  example: "CGF01"
                tanggal:
                  type: string
                  format: date
                  example: "2024-01-15"
                records:
                  type: array
                  items:
                    $ref: '#/components/schemas/AttendanceRecord'
                  example:
                    - no_jemaat: 1
                      keterangan: "hadir"
                    - no_jemaat: 2
                      keterangan: "izin"
      responses:
        '201':
          description: Bulk attendance recorded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Attendance'
                  message:
                    type: string
                    example: "Attendance records created successfully"
  /attendance/{id}:
    parameters:
      - name: id
        in: path
        required: true
        description: Attendance record ID
        schema:
          type: integer

    get:
      tags: [Attendance]
      summary: Get attendance record by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Attendance]
      summary: Update attendance record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttendanceUpdate'
      responses:
        '200':
          description: Attendance updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AttendanceResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Attendance]
      summary: Delete attendance record
      responses:
        '200':
          description: Attendance deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== MINISTRY INFO ENDPOINTS ====================
...
  /events:
    get:
      tags: [Events]
      summary: Get all events
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: search
          in: query
          description: Search by event name
          schema:
            type: string
        - name: category
          in: query
          description: Filter by event category
          schema:
            type: string
            enum: [Camp, Retreat, Quarterly, Monthly, Special]
        - name: start_date
          in: query
          description: Start date filter
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: End date filter
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventListResponse'
              example:
                success: true
                data:
                  - event_id: 1
                    event_name: "Annual Youth Camp"
                    event_date: "2024-03-15"
                    category: "Camp"
                    location: "Bogor Retreat Center"
                    description: "Annual youth camping event"
                pagination:
                  page: 1
                  limit: 10
                  total: 20
                  totalPages: 2

    post:
      tags: [Events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventCreate'
      responses:
        '201':
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /events/{event_id}:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer

    get:
      tags: [Events]
      summary: Get event by ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Events]
      summary: Update event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventUpdate'
      responses:
        '200':
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Events]
      summary: Delete event
      responses:
        '200':
          description: Event deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== EVENT PARTICIPATION ENDPOINTS ====================
  /events/{event_id}/participants:
    parameters:
      - name: event_id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Events]
      summary: Get event participants
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: role
          in: query
          description: Filter by role
          schema:
            type: string
            enum: [Peserta, Panitia, Volunteer]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventParticipationListResponse'

    post:
      tags: [Events]
      summary: Register event participant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventParticipationCreate'
      responses:
        '201':
          description: Participant registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventParticipationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /members/{no_jemaat}/events:
    parameters:
      - $ref: '#/components/parameters/MemberId'
    get:
      tags: [Events, Members]
      summary: Get events for a member
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventParticipationListResponse'

  # ==================== ANALYTICS ENDPOINTS ====================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ==================== PARAMETERS ====================
  parameters:
    Page:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    Limit:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...
    MemberId:
      name: no_jemaat
      in: path
      required: true
      description: Member number (primary key)
      schema:
        type: integer
        minimum: 1

  # ==================== RESPONSES ====================
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 400
              message: "Validation error"
              details:
                - field: "nama_jemaat"
                  message: "Name is required"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: 404
              message: "Resource not found"
...
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: integer
              example: 400
            message:
              type: string
              example: "Error message"
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10

    # Member Schemas
    Member:
      type: object
      properties:
        no_jemaat:
          type: integer
          description: Member number (primary key)
          example: 1
        nama_jemaat:
          type: string
          example: "John Doe"
        jenis_kelamin:
          type: string
          enum: [Laki-laki, Perempuan]
          example: "Laki-laki"
        tanggal_lahir:
          type: string
          format: date
          example: "1990-05-15"
        tahun_lahir:
          type: integer
          example: 1990
        bulan_lahir:
          type: integer
          minimum: 1
          maximum: 12
          example: 5
        kuliah_kerja:
          type: string
          example: "Software Engineer"
        no_handphone:
          type: string
          example: "081234567890"
        ketertarikan_cgf:
          type: string
          example: "Yes"
        nama_cgf:
          type: string
          example: "Grace CGF"
        kategori_domisili:
          type: string
          example: "Jakarta"
        alamat_domisili:
          type: string
          example: "Jl. Sudirman No. 123"
...
    Attendance:
      type: object
      properties:
        id:
          type: integer
          description: Auto-generated ID
          example: 1
        no_jemaat:
          type: integer
          example: 1
        cg_id:
          type: string
          example: "CGF01"
        tanggal:
          type: string
          format: date
          example: "2024-01-15"
        keterangan:
          type: string
          enum: [hadir, izin, tidak_hadir, tamu]
          example: "hadir"
        member:
          $ref: '#/components/schemas/Member'
    AttendanceCreate:
      type: object
      required: [no_jemaat, cg_id, tanggal, keterangan]
      properties:
        no_jemaat:
          type: integer
        cg_id:
          type: string
        tanggal:
          type: string
          format: date
        keterangan:
          type: string
          enum: [hadir, izin, tidak_hadir, tamu]
    AttendanceRecord:
      type: object
      required: [no_jemaat, keterangan]
      properties:
        no_jemaat:
          type: integer
          example: 1
        keterangan:
          type: string
          enum: [hadir, izin, tidak_hadir, tamu]
    AttendanceUpdate:
      type: object
      properties:
        keterangan:
          type: string
          enum: [hadir, izin, tidak_hadir, tamu]
        tanggal:
          type: string
          format: date
    AttendanceResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Attendance'
    AttendanceListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Attendance'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Ministry Schemas
...
    Event:
      type: object
      properties:
        event_id:
          type: integer
          example: 1
        event_name:
          type: string
          example: "Annual Youth Camp"
        event_date:
          type: string
          format: date
          example: "2024-03-15"
        category:
          type: string
          enum: [Camp, Retreat, Quarterly, Monthly, Special]
          example: "Camp"
        location:
          type: string
          example: "Bogor Retreat Center"
        description:
          type: string
          example: "Annual youth camping event"
    EventCreate:
      type: object
      required: [event_name, event_date, category]
      properties:
        event_name:
          type: string
        event_date:
          type: string
          format: date
        category:
          type: string
          enum: [Camp, Retreat, Quarterly, Monthly, Special]
        location:
          type: string
        description:
          type: string
    EventUpdate:
      type: object
      required: [event_name, event_date, category]
      properties:
        event_name:
          type: string
        event_date:
          type: string
          format: date
        category:
          type: string
          enum: [Camp, Retreat, Quarterly, Monthly, Special]
        location:
          type: string
        description:
          type: string
    EventResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/Event'
    EventListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        pagination:
          $ref: '#/components/schemas/Pagination'

    # Event Participation Schemas
    EventParticipation:
      type: object
      properties:
        id:
          type: integer
          example: 1
        event_id:
          type: integer
          example: 1
        no_jemaat:
          type: integer
          example: 1
        role:
          type: string
          enum: [Peserta, Panitia, Volunteer]
          example: "Peserta"
        registered_at:
          type: string
          format: date-time
          example: "2024-01-20T09:00:00Z"
        event:
          $ref: '#/components/schemas/Event'
        member:
          $ref: '#/components/schemas/Member'
    EventParticipationCreate:
      type: object
      required: [event_id, no_jemaat, role]
      properties:
        event_id:
          type: integer
        no_jemaat:
          type: integer
        role:
          type: string
          enum: [Peserta, Panitia, Volunteer]
...
    EventParticipationResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/EventParticipation'
    EventParticipationListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/EventParticipation'
        pagination:
          $ref: '#/components/schemas/Pagination'